git:toc
## Learning by doing 

The best way to learn to code is to code. With tools like gitx, you can organize your code by commits. There’s so much learning in a specific commit message that I started wishing I could just convert commits into blog posts. But the problem is I have to context switch, reformat, and maintain a separate project. So, that gave me an idea. What if I could have a table of contents that organizes my commits based on specific topics. 

## Write messages instead of posts

Commit messages have space for more than a one line title. I never really thought of using this space beyond a single line and link to a jira ticket. But what if this is the section to write a post? So, here I am writing a post in the commit message. 

## Commits as comments

Another benefit of using commits as posts is that fellow developers can ask questions and comment on a specific line of the commit.

## git:toc

Imagine a set gem that helps you organize your commit history and then publish a helpful table of contents. 

### Example Output

Learning Ruby
I had wanted to learn more about specific methods
07b9657 The difference between #map and #select

Setting up Capybara
I want to get do end to end testing
1fdb874 Trying to get Capybara’s #page method, but not successful
35969a0 Add minitest & capybara via minitest-rails-capybara
0b040c2 Capybara works, yay!


### config/gittoc.yml

The config sets the title, summary, and tags that cluster commits under this topic.

### rake tasks

The first task is trying for only a single tag, and filters out the merges. It also is reverse order the way a normal Table of contents works. 

I envision a `git:publish` that will convert info into a `index.html` page

### What do you think?